requirements.txt
Click==7.0
Jinja2==2.10
MarkupSafe==1.1.0
Unidecode==1.0.23
Werkzeug==0.14.1
certifi==2018.11.29
chardet==3.0.4
feedparser==5.2.1
idna==2.8
itsdangerous==1.1.0
markovify==0.7.1
mecab-python3==0.996.1
oauthlib==2.1.0
requests==2.21.0
requests-oauthlib==1.0.0
urllib3==1.24.1
